+++ crates/sncast/src/lib.rs
@@ -28,8 +28,7 @@ use crate::response::errors::SNCastProviderError;
 use conversions::serde::serialize::CairoSerialize;
 use serde::de::DeserializeOwned;
 use shared::rpc::create_rpc_client;
-use starknet::accounts::AccountFactoryError;
-use starknet::signers::local_wallet::SignError;
+use starknet::accounts::{AccountFactory, AccountFactoryError};
 use std::collections::HashMap;
 use std::thread::sleep;
 use std::time::Duration;
@@ -542,10 +541,13 @@ pub fn handle_rpc_error(error: impl Into<SNCastProviderError>) -> Error {
 }
 
 #[must_use]
-pub fn handle_account_factory_error(err: AccountFactoryError<SignError>) -> anyhow::Error {
+pub fn handle_account_factory_error<T>(err: AccountFactoryError<T::SignError>) -> anyhow::Error
+where
+    T: AccountFactory + Sync,
+{
     match err {
         AccountFactoryError::Provider(error) => handle_rpc_error(error),
-        error => anyhow!(error),
+        error => anyhow!(error.to_string()),
     }
 }