+++ crates/sncast/tests/e2e/account/create.rs
@@ -1,16 +1,23 @@
-use crate::helpers::constants::{DEVNET_OZ_CLASS_HASH_CAIRO_0, URL};
+use crate::helpers::constants::{
+    ARGENT_ACCOUNT_CLASS_HASH, DEVNET_OZ_CLASS_HASH_CAIRO_0, DEVNET_OZ_CLASS_HASH_CAIRO_1, URL,
+};
 use crate::helpers::fixtures::{copy_file, default_cli_args};
 use crate::helpers::runner::runner;
 use configuration::copy_config_to_tempdir;
 use indoc::indoc;
 
-use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains, AsOutput};
+use serde_json::{json, to_string_pretty};
+use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
+use snapbox::assert_matches;
 use sncast::helpers::constants::CREATE_KEYSTORE_PASSWORD_ENV_VAR;
 use std::{env, fs};
 use tempfile::tempdir;
+use test_case::test_case;
 
+#[test_case("oz"; "oz_account_type")]
+#[test_case("argent"; "argent_account_type")]
 #[tokio::test]
-pub async fn test_happy_case() {
+pub async fn test_happy_case(account_type: &str) {
     let temp_dir = tempdir().expect("Unable to create a temporary directory");
     let accounts_file = "accounts.json";
 
@@ -22,36 +29,48 @@ pub async fn test_happy_case() {
         "account",
         "create",
         "--name",
-        "my_account_create_happy",
+        "my_account",
         "--salt",
         "0x1",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        account_type,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
-    let output = snapbox.assert();
+    let output = snapbox.assert().success();
 
-    let stdout_str = output.as_stdout();
-    assert!(stdout_str.contains("command: account create"));
-    assert!(stdout_str.contains("max_fee: "));
-    assert!(!stdout_str.contains("max_fee: 0x"));
-    assert!(stdout_str.contains("address: 0x"));
-    assert!(stdout_str.contains(
-        "add_profile: --add-profile flag was not set. No profile added to snfoundry.toml"
-    ));
+    assert_stdout_contains(
+        output,
+        indoc! {r"
+        command: account create
+        add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
+        address: 0x[..]
+        max_fee: [..]
+        message: Account successfully created. Prefund generated address with at least <max_fee> tokens. It is good to send more in the case of higher demand.
+        "},
+    );
 
     let contents = fs::read_to_string(temp_dir.path().join(accounts_file))
         .expect("Unable to read created file");
-    assert!(contents.contains("my_account"));
-    assert!(contents.contains("alpha-sepolia"));
-    assert!(contents.contains("private_key"));
-    assert!(contents.contains("public_key"));
-    assert!(contents.contains("address"));
-    assert!(contents.contains("salt"));
-    assert!(contents.contains("class_hash"));
-    assert!(contents.contains("legacy"));
-    assert!(contents.contains("type"));
+
+    let expected = json!(
+        {
+            "alpha-sepolia": {
+                "my_account": {
+                    "address": "0x[..]",
+                    "class_hash": "0x[..]",
+                    "deployed": false,
+                    "legacy": false,
+                    "private_key": "0x[..]",
+                    "public_key": "0x[..]",
+                    "salt": "0x1",
+                    "type": get_formatted_account_type(account_type)
+                }
+            }
+        }
+    );
+
+    assert_matches(to_string_pretty(&expected).unwrap(), contents);
 }
 
 #[tokio::test]
@@ -66,6 +85,8 @@ pub async fn test_invalid_class_hash() {
         accounts_file,
         "account",
         "create",
+        "--type",
+        "oz",
         "--class-hash",
         "0x10101",
         "--name",
@@ -102,6 +123,8 @@ pub async fn test_happy_case_generate_salt() {
         "my_account",
         "--class-hash",
         DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        "oz",
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -143,8 +166,6 @@ pub async fn test_happy_case_add_profile() {
         "my_account",
         "--add-profile",
         "my_account",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -181,8 +202,6 @@ pub async fn test_happy_case_accounts_file_already_exists() {
         "my_account",
         "--salt",
         "0x1",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -218,8 +237,6 @@ pub async fn test_profile_already_exists() {
         "myprofile",
         "--add-profile",
         "default",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -238,14 +255,7 @@ pub async fn test_profile_already_exists() {
 pub async fn test_account_already_exists() {
     let mut args = default_cli_args();
     args.append(&mut vec![
-        "account",
-        "create",
-        "--name",
-        "user1",
-        "--salt",
-        "0x1",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "account", "create", "--name", "user1", "--salt", "0x1",
     ]);
 
     let snapbox = runner(&args);
@@ -260,8 +270,10 @@ pub async fn test_account_already_exists() {
     );
 }
 
+#[test_case("oz"; "oz_account_type")]
+#[test_case("argent"; "argent_account_type")]
 #[tokio::test]
-pub async fn test_happy_case_keystore() {
+pub async fn test_happy_case_keystore(account_type: &str) {
     let temp_dir = tempdir().expect("Unable to create a temporary directory");
     let keystore_file = "my_key.json";
     let account_file = "my_account.json";
@@ -276,8 +288,8 @@ pub async fn test_happy_case_keystore() {
         account_file,
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        account_type,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -290,12 +302,12 @@ pub async fn test_happy_case_keystore() {
         message: Account successfully created[..]
     "});
 
+    assert!(temp_dir.path().join(keystore_file).exists());
+
     let contents = fs::read_to_string(temp_dir.path().join(account_file))
         .expect("Unable to read created file");
-    assert!(contents.contains("\"deployment\": {"));
-    assert!(contents.contains("\"variant\": {"));
-    assert!(contents.contains("\"version\": 1"));
-    assert!(contents.contains("\"legacy\": true"));
+
+    assert_matches(get_keystore_account_pattern(account_type, None), contents);
 }
 
 #[tokio::test]
@@ -317,8 +329,6 @@ pub async fn test_happy_case_keystore_add_profile() {
         account_file,
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
         "--add-profile",
         "with_keystore",
     ];
@@ -335,7 +345,7 @@ pub async fn test_happy_case_keystore_add_profile() {
     assert!(contents.contains("\"deployment\": {"));
     assert!(contents.contains("\"variant\": {"));
     assert!(contents.contains("\"version\": 1"));
-    assert!(contents.contains("\"legacy\": true"));
+    assert!(contents.contains("\"legacy\": false"));
 
     let contents = fs::read_to_string(tempdir.path().join("snfoundry.toml"))
         .expect("Unable to read snfoundry.toml");
@@ -360,8 +370,6 @@ pub async fn test_keystore_without_account() {
         keystore_file,
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -398,8 +406,6 @@ pub async fn test_keystore_file_already_exists() {
         account_file,
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -437,8 +443,6 @@ pub async fn test_keystore_account_file_already_exists() {
         account_file,
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -473,6 +477,8 @@ pub async fn test_happy_case_keystore_int_format() {
         "create",
         "--class-hash",
         DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        "oz",
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -511,8 +517,6 @@ pub async fn test_happy_case_keystore_hex_format() {
         "--hex-format",
         "account",
         "create",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -530,5 +534,56 @@ pub async fn test_happy_case_keystore_hex_format() {
     assert!(contents.contains("\"deployment\": {"));
     assert!(contents.contains("\"variant\": {"));
     assert!(contents.contains("\"version\": 1"));
-    assert!(contents.contains("\"legacy\": true"));
+    assert!(contents.contains("\"legacy\": false"));
+}
+
+fn get_formatted_account_type(account_type: &str) -> &str {
+    match account_type {
+        "oz" => "open_zeppelin",
+        _ => account_type,
+    }
+}
+
+fn get_keystore_account_pattern(account_type: &str, class_hash: Option<&str>) -> String {
+    let account_json = match account_type {
+        "oz" => {
+            json!(
+                {
+                    "version": 1,
+                    "variant": {
+                        "type": "open_zeppelin",
+                        "version": 1,
+                        "public_key": "0x[..]",
+                        "legacy": false,
+                    },
+                    "deployment": {
+                        "status": "undeployed",
+                        "class_hash": class_hash.unwrap_or(DEVNET_OZ_CLASS_HASH_CAIRO_1),
+                        "salt": "0x[..]",
+                    }
+                }
+            )
+        }
+        "argent" => {
+            json!(
+                {
+                    "version": 1,
+                    "variant": {
+                        "type": "argent",
+                        "version": 1,
+                        "owner": "0x[..]",
+                        "guardian": "0x0"
+                    },
+                    "deployment": {
+                        "status": "undeployed",
+                        "class_hash": class_hash.unwrap_or(ARGENT_ACCOUNT_CLASS_HASH),
+                        "salt": "0x[..]",
+                    }
+                }
+            )
+        }
+        _ => panic!("Incorrect account type"),
+    };
+
+    to_string_pretty(&account_json).unwrap()
 }